debt - move actions around to code editor
